First Unit Test

+++ src/main/kotlin/com/example/calculator/CalculatorService.kt
@@ -4,5 +4,5 @@ import org.springframework.stereotype.Service
 
 @Service
 class CalculatorService {
-    fun add(a: Int, b: Int): Int = a + b
+    fun add(a: Long, b: Long): Long = a + b
 }
